# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: BridgeSync API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /dept/{deptName}/applications:
    get:
      operationId: getDeptApplications
      tags: [departments]
      summary: Get officer queue for a department
      parameters:
        - name: deptName
          in: path
          required: true
          schema:
            type: string
        - name: officerId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of dept applications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DeptApplication"
  /dept/{deptName}/applications/{appId}/status:
    put:
      operationId: updateDeptApplicationStatus
      tags: [departments]
      summary: Officer updates application status
      parameters:
        - name: deptName
          in: path
          required: true
          schema:
            type: string
        - name: appId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStatusBody"
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeptApplication"
  /dept/{deptName}/applications/{appId}/request-document:
    post:
      operationId: requestDocument
      tags: [departments]
      summary: Officer requests a document from citizen
      parameters:
        - name: deptName
          in: path
          required: true
          schema:
            type: string
        - name: appId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestDocumentBody"
      responses:
        "201":
          description: Document requested
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentRequest"
  /officers/{deptName}:
    get:
      operationId: getOfficers
      tags: [departments]
      summary: Get officers for a department
      parameters:
        - name: deptName
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of officers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Officer"
...
  /departments:
    get:
      operationId: getDepartments
      tags: [departments]
      summary: Get all departments
      responses:
        "200":
          description: Departments list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Department"
    post:
      operationId: onboardDepartment
      tags: [departments]
      summary: Onboard a new department
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OnboardDepartmentBody"
      responses:
        "201":
          description: Department onboarded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
components:
  schemas:
...
    DeptApplication:
      type: object
      properties:
        id:
          type: integer
        appId:
          type: string
        department:
          type: string
        nativeFieldsJson:
          type: string
        status:
          type: string
        assignedOfficer:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
        citizenName:
          type: ["string", "null"]
        businessType:
          type: ["string", "null"]
        district:
          type: ["string", "null"]
        daysPending:
          type: ["integer", "null"]
      required:
        - id
        - appId
        - department
        - status
        - createdAt
        - updatedAt
    DocumentRequest:
      type: object
      properties:
        id:
          type: integer
        appId:
          type: string
        department:
          type: string
        documentName:
          type: string
        requestedAt:
          type: string
        fulfilledAt:
          type: ["string", "null"]
        citizenResponse:
          type: ["string", "null"]
      required:
        - id
        - appId
        - department
        - documentName
        - requestedAt
...
    UpdateStatusBody:
      type: object
      properties:
        status:
          type: string
        actor:
          type: string
        notes:
          type: string
      required:
        - status
    RequestDocumentBody:
      type: object
      properties:
        documentName:
          type: string
        officerName:
          type: string
      required:
        - documentName
...
    Officer:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        department:
          type: string
        role:
          type: string
      required:
        - id
        - name
        - department
        - role
...
    Department:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        systemType:
          type: string
        adapterStatus:
          type: string
        onboardedAt:
          type: ["string", "null"]
      required:
        - id
        - name
        - systemType
        - adapterStatus
    OnboardDepartmentBody:
      type: object
      properties:
        name:
          type: string
        systemType:
          type: string
      required:
        - name
        - systemType