# lib/api-zod/src/generated/types/onboardDepartmentBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BridgeSync API specification
 * OpenAPI spec version: 0.1.0
 */

export interface OnboardDepartmentBody {
  name: string;
  systemType: string;
}